Add parser for complex options
Instead of only allowing boolean and with some exceptions numeric
values as option value, complex options are parsed like a simple
version of JSON. They are needed in order to define values in groups
and with this to get things like #28 work properly.
Here is the parser implementation. It parses only, if the option is
inside /*jshint blocks. The values of the inner options can be
boolean, string, numeric or complex. Single identifiers are treated
as strings, so
   { x: function } === { x: "function" }

If the option name is not known, a warning is generated. The values
will be parsed, but ignored.